% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_piper_ggplot.R
\name{blank_ggplot_piper}
\alias{blank_ggplot_piper}
\title{Plot a blank Piper diagram with ggplot.}
\usage{
blank_ggplot_piper()
}
\value{
A blank ggplot Piper diagram.
}
\description{
Plot a blank Piper diagram with ggplot.
}
\examples{
blank_ggplot_piper()
}
