% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.R
\name{cdf}
\alias{cdf}
\title{Cumulative distribution function (CDF)}
\usage{
cdf(var, unique = FALSE)
}
\arguments{
\item{var}{A numeric vector.}

\item{unique}{A boolean for specifying if returning only unique values.}
}
\value{
A list with two numeric vectors: 1. Probability exceedance and 2. Associated var value
}
\description{
Cumulative distribution function (CDF)
}
\examples{
data(KarstMod_dataset)
cdf(KarstMod_dataset$Qobs)
}
