% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_consec.R
\name{date_consec}
\alias{date_consec}
\title{Identify consecutive dates.}
\usage{
date_consec(date)
}
\arguments{
\item{date}{A date vector.}
}
\value{
A dataframe.
}
\description{
Identify consecutive dates.
}
\examples{
date <- as.Date(c("2020-01-01", "2020-01-02", "2020-01-03", "2020-02-05", "2020-02-06", "2020-07-07"))
km_import(path = path)
}
