% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_year.R
\name{density_year}
\alias{density_year}
\title{Generate plot to show the density of criterion score by hydrological year.}
\usage{
density_year(
  date,
  sim,
  obs,
  interval,
  month_hyear = 9,
  crit = c("NSE", "KGE", "rpearson", "rspearman", "KGE_m", "KGE_m2", "KGENP", "LME",
    "LCE"),
  sf = c(1, 1, 1),
  bar_width = 0.048,
  palette = c("#eff3ff", "#2473B6")
)
}
\arguments{
\item{date}{A date vector.}

\item{sim}{A numeric vector of simulated time series.}

\item{obs}{A numeric vector of observed time series.}

\item{interval}{A numeric value corresponding to the desired class interval.}

\item{month_hyear}{A numeric value corresponding to the first month of the hydrological year.}

\item{crit}{A string value corresponding to the desired performance criterion for the evaluation.}

\item{sf}{A numeric vector for specifying scaling factors if needed for correlation, variability and bias, respectively. Default values are 1-1-1.}

\item{bar_width}{A numeric value for specifying the width of the bar.}

\item{palette}{A two string vector with the desired extreme colors for the palette.}
}
\value{
A plot. Best export dimensions for default width are width = 8, height = 5.
}
\description{
Generate plot to show the density of criterion score by hydrological year.
}
\examples{
data(KarstMod_dataset)
density_year(KarstMod_dataset$date, KarstMod_dataset$QobsS + runif(365, -2, +10), KarstMod_dataset$QobsS, interval = 10)
}
