% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_KarstMod.R
\name{import_KarstMod}
\alias{import_KarstMod}
\title{Import a dataset in KarstMod input format}
\usage{
import_KarstMod(path, version = c(3, 2))
}
\arguments{
\item{path}{Path to file.}

\item{version}{KarstMod version.}
}
\value{
A dataframe.
}
\description{
Import a dataset in KarstMod input format
}
\examples{
path <- system.file("extdata", "KarstMod_dataset.txt", package = "hydros1mple")
import_KarstMod(path = path)
}
