% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_km.R
\name{import_km}
\alias{import_km}
\title{Import a dataset in KarstMod input format}
\usage{
import_km(path, version = c("3", "2"))
}
\arguments{
\item{path}{Path to file.}

\item{version}{KarstMod version.}
}
\value{
A dataframe.
}
\description{
Import a dataset in KarstMod input format
}
\examples{
path <- system.file("extdata", "KarstMod_dataset.txt", package = "hydros1mple")
import_km(path = path)
}
