% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight.R
\name{insight}
\alias{insight}
\title{Helping function for analysing sim/obs discharge from hydrological models.}
\usage{
insight(date, sim, obs, na.rm = FALSE, interactive = FALSE)
}
\arguments{
\item{date}{A date vector.}

\item{sim}{A numeric vector of simulated time series.}

\item{obs}{A numeric vector of observed time series.}

\item{na.rm}{A boolean for na remove.}

\item{interactive}{A boolean for interactive plot (ggplotly). Only for 1. Daily hydrograph}
}
\value{
A list with 6 items comparing sim and obs:
\itemize{
\item daily_hydrograph: A plot of daily discharge (can be interactive)
\item performance_criteria: A list of different performance criteria (see hydros1mple::score)
\item monthly_hydrograph: A plot of mean monthly discharge
\item yearly_hydrograph: A plot of yearly discharge
\item boxplot: A boxplot of the distribution of discharge
\item cdf: The cumulative distribution function of discharge
}
}
\description{
Helping function for analysing sim/obs discharge from hydrological models.
}
\examples{
data(KarstMod_dataset)
insight(KarstMod_dataset$date, KarstMod_dataset$QobsS + runif(365, -2, +10), KarstMod_dataset$QobsS)
}
