% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KarstMod_dataset.R
\docType{data}
\name{KarstMod_dataset}
\alias{KarstMod_dataset}
\title{A KarstMod dataset.}
\format{
A data frame with 365 rows and 7 variables:
\describe{
\item{date}{date}
\item{comment}{row number}
\item{P}{precipitation}
\item{T}{temperature}
\item{RH}{relative humidity}
\item{ET}{evapotranspiration}
\item{QobsS}{observed discharge}
...
}
}
\usage{
KarstMod_dataset
}
\description{
A KarstMod dataset.
}
\keyword{datasets}
