% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_import.R
\name{km_import}
\alias{km_import}
\title{Import a dataset in KarstMod input format}
\usage{
km_import(path, version = c("3", "2"))
}
\arguments{
\item{path}{Path to file.}

\item{version}{KarstMod version.}
}
\value{
A dataframe.
}
\description{
Import a dataset in KarstMod input format
}
\examples{
path <- system.file("extdata", "KarstMod_dataset.txt", package = "hydros1mple")
km_import(path = path)
}
