% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_properties.R
\name{km_properties}
\alias{km_properties}
\title{Helpfer function to extract information from KarstMod .properties file}
\usage{
km_properties(path)
}
\arguments{
\item{path}{A string vector corresponding to the path of the .properties file.}
}
\value{
A list with different information extracted from the file.
}
\description{
Helpfer function to extract information from KarstMod .properties file
}
\examples{
# km_properties("C:/users/karstmod/model.properties")
}
