% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meq_to_percent.R
\name{meq_to_percent}
\alias{meq_to_percent}
\title{Convert analyses in meq/L to percent for Piper diagram.}
\usage{
meq_to_percent(
  data,
  cations = c("Ca", "Mg", "Na", "K"),
  anions = c("Cl", "SO4", "HCO3", "CO3")
)
}
\arguments{
\item{data}{A dataframe with ions as columns.}

\item{cations}{A string vector of cations.}

\item{anions}{A string vector of anions.}
}
\value{
A dataframe with a new columns of each ion in percent.
}
\description{
Convert analyses in meq/L to percent for Piper diagram.
}
\examples{
data(pc_data)

# First convert to meq/L
pc_data_meq <- mgl_to_meq(pc_data, ions = c("Ca", "Mg", "Na", "K", "SO4", "Cl", "HCO3"))

# Define anions argument to exclude "CO3" which is absent in this dataset
meq_to_percent(pc_data_meq, anions = c("Cl", "SO4", "HCO3"))
}
