% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{PET}
\alias{PET}
\title{Calculation of Potential Evapotranspiration (PET) with different methods.}
\usage{
PET(
  date,
  t,
  method = c("oudin2005", "hargreaves1985", "turc1961"),
  tmax = NULL,
  tmin = NULL,
  rh = NULL,
  latitude = NULL,
  krs = 0.17,
  Rs = NULL,
  k1 = 100,
  k2 = 5
)
}
\arguments{
\item{date}{A date vector.}

\item{t}{A numeric vector.}

\item{method}{A string value. Either "oudin2005", "hargreaves1985" or "turc1961".}

\item{tmax}{A numeric vector.}

\item{tmin}{A numeric vector.}

\item{rh}{A numeric vector.}

\item{latitude}{A numeric value. Latitude (in degrees) of the meteorological station. This is for estimating Ra.}

\item{krs}{A numeric value. Radiation adjustment coefficient (generally between 0.16 and 0.19).}

\item{Rs}{A numeric vector.}

\item{k1}{A numeric value. Used for Oudin's method.}

\item{k2}{A numeric value. Used for Oudin's method.}
}
\value{
A numeric vector.
}
\description{
Calculation of Potential Evapotranspiration (PET) with different methods.
}
\examples{
data(KarstMod_dataset)
PET(date = KarstMod_dataset$date, t = KarstMod_dataset$T, latitude = 48)
}
