% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1v.R
\name{plot1v}
\alias{plot1v}
\title{Plot one variable.}
\usage{
plot1v(
  date,
  data,
  var = c("q", "p", "t", "et"),
  group = c(NA, "hour", "day", "month", "year"),
  type = c("mean", "sum"),
  xlab = waiver(),
  ylab = waiver(),
  title = waiver(),
  date_breaks = waiver(),
  date_labels = waiver()
)
}
\arguments{
\item{date}{A date vector.}

\item{data}{A numeric vector.}

\item{var}{A string value. Either "q" for discharge, "p" for precipitation, "t" for temperature, "et" for evapotranspiration.}

\item{group}{A string value for grouping data. Either NA for no group, or "hour", "day", "month" or "year".}

\item{type}{A string value for aggregate type. Either "mean" or "sum".}

\item{xlab}{A string value.}

\item{ylab}{A string value.}

\item{title}{A string value.}

\item{date_breaks}{A string giving the formatting specification for the labels. Codes are defined in strftime().}

\item{date_labels}{A string giving the distance between breaks like "2 weeks", or "10 years". If both breaks and date_breaks are specified, date_breaks wins.}
}
\value{
A plot.
}
\description{
Plot one variable.
}
\examples{
data(KarstMod_dataset)

# Basic hydrograph

plot1v(KarstMod_dataset$date, KarstMod_dataset$QobsS, var = "q")

# Add details

plot1v(KarstMod_dataset$date, KarstMod_dataset$QobsS, var = "q",
       xlab = "date", ylab = "discharge", title = "Hydrograph",
       date_breaks = "3 month", date_labels = "\%Y-\%m")
}
