% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2qp.R
\name{plot2qp}
\alias{plot2qp}
\title{Plot discharge and precipitation on the same graph.}
\usage{
plot2qp(
  date,
  q,
  p,
  maxRange = 20,
  coeff = 1,
  precip_breaks = c(0, 25, 50, 75, 100, 125, 150),
  xlab = waiver(),
  ylab = "Discharge",
  ylab2 = "Precipitation",
  title = waiver(),
  legend = c("none", "bottom", "right", "top", "left"),
  date_breaks = waiver(),
  date_labels = waiver()
)
}
\arguments{
\item{date}{A date vector.}

\item{q}{A numeric vector for discharge.}

\item{p}{A numeric vector for precipitation.}

\item{maxRange}{A numeric value. Set the width of the discharge axis.}

\item{coeff}{A numeric value. Set the shrink coefficient of precipitation (visual ratio).}

\item{precip_breaks}{A numeric vector. Set the axis ticks values. Manual breaks allows to avoid truncated value caused by negative margin.}

\item{xlab}{A string value.}

\item{ylab}{A string value.}

\item{ylab2}{A string value.}

\item{title}{A string value.}

\item{legend}{A string value for the position of the legend. Either "none" for no legend, or "bottom", "right", "top", "left".}

\item{date_breaks}{A string giving the formatting specification for the labels. Codes are defined in strftime().}

\item{date_labels}{A string giving the distance between breaks like "2 weeks", or "10 years". If both breaks and date_breaks are specified, date_breaks wins.}
}
\value{
A plot with discharge and precipitation.
}
\description{
Plot discharge and precipitation on the same graph.
}
\examples{
data(KarstMod_dataset)

plot2qp(KarstMod_dataset$date, KarstMod_dataset$QobsS, KarstMod_dataset$P,
        maxRange = 150, coef = 1.5)
}
