% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq.R
\name{plot_freq}
\alias{plot_freq}
\title{Frequency barplot of variables}
\usage{
plot_freq(
  var_list,
  interval,
  x_name,
  basesize = 16,
  height = 0.3,
  color_low = "lightyellow",
  color_high = "darkred",
  legend_name = "Freq",
  rect_width = 0.1
)
}
\arguments{
\item{var_list}{A list of named vector.}

\item{interval}{An integer corresponding to the grouping interval.}

\item{x_name}{A string for the x-axis name.}

\item{color_low}{A color for small frequencies.}

\item{color_high}{A color for high frequencies.}

\item{legend_name}{A string for the name of the legend.}

\item{rect_width}{A value for the linewidth of the rectangle outline.}
}
\value{
A plot.
}
\description{
Warning: at the moment, only works for variables with min value greater than 10
}
\examples{
plot_freq(var_list = c("site_name" = c(1, 2, 4, 5, 6, 5, 5, 5), "site_name2" = c(2, 3, 5, 10, 15, 15, 15),
          interval = 2, x_name = "var")
}
