% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SR.R
\name{run_snow_routine}
\alias{run_snow_routine}
\title{Run the snow routine with P and T data}
\usage{
run_snow_routine(precip, temp, param, subcatchment_table = NULL)
}
\arguments{
\item{precip}{A numeric vector for precipitation.}

\item{temp}{A numeric vector for temperature}

\item{param}{A numeric vector for model parameters: 1. Ts = threshold temperature \link{C}; 2. MF = melt factor \link{mm/C}; 3. CFR = refreezing factor \link{-}; 4. CWHv= Water holding capacity of snow \link{-}}

\item{subcatchment_table}{A file path to a subctachment table.}
}
\value{
A vector of the output of the snow routine (Psr).
}
\description{
Run the snow routine with P and T data
}
