% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_color.R
\name{sample_color}
\alias{sample_color}
\title{Random sample of colors with/without grey scale}
\usage{
sample_color(n, grey = FALSE)
}
\arguments{
\item{n}{A number. Number of unique colors desired.}

\item{grey}{A boolean. FALSE remove all grey-like colors from the pool.}
}
\value{
A string vector of color names.
}
\description{
Random sample of colors with/without grey scale
}
\examples{
sample_color(5)
sample_color(5, grey = TRUE)
}
