% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Calculate performance criteria scores between a simulated and observed time series}
\usage{
score(
  sim,
  obs,
  crit = c("NSE", "BE", "C2M", "rpearson", "rspearman", "RMSE", "KGE", "KGE_m", "KGE_m2",
    "KGENP", "LME", "LCE", "KGE_abs", "KGE_m_abs", "KGENP_abs", "bias", "MAE", "beta",
    "beta_abs", "alpha", "gamma", "beta_n", "CI"),
  sf = c(1, 1, 1),
  na.rm = FALSE,
  allow_NA = FALSE,
  format = c("vector", "list")
)
}
\arguments{
\item{sim}{A numeric vector of simulated time series.}

\item{obs}{A numeric vector of observed time series.}

\item{crit}{A string vector of performance criteria to calculate.}

\item{sf}{A numeric vector for specifying scaling factors if needed for correlation, variability and bias, respectively. Default values are 1-1-1.}

\item{na.rm}{A boolean for na remove.}

\item{format}{A string vector for defining the output format, either vector or list.}
}
\value{
A numeric named vector or list of scores.
}
\description{
Calculate performance criteria scores between a simulated and observed time series
}
\examples{
data(KarstMod_dataset)
obs <- KarstMod_dataset$QobsS
sim <- obs + runif(nrow(KarstMod_dataset), 0, sd(obs)) # Synthetic sim

score(sim, obs, crit = c("NSE", "KGE", "KGE_m"), sf = c(2, 1, 1), na.rm = FALSE)
}
