% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_seq.R
\name{score_seq}
\alias{score_seq}
\title{Generate plot to evaluate a model by hydrological year on different subregimes}
\usage{
score_seq(
  date,
  sim,
  obs,
  freq = c("hyear", "month"),
  month_hyear = 9,
  lf_quantile = 0.1,
  hf_quantile = 0.9,
  crit = c("NSE", "KGE", "rpearson", "rspearman", "KGE_m", "KGE_m2", "KGENP", "LME",
    "LCE"),
  sf = c(1, 1, 1),
  rel_heights = c(1, 0.9, 1.55)
)
}
\arguments{
\item{date}{A date vector.}

\item{sim}{A numeric vector of simulated time series.}

\item{obs}{A numeric vector of observed time series.}

\item{freq}{A string specifying the frequence of analysis, either "hyear" for hydrological year or "month".}

\item{month_hyear}{A numeric value corresponding to the first month of the hydrological year.}

\item{lf_quantile}{A numeric value corresponding to the low flow threshold (lower than).}

\item{hf_quantile}{A numeric value corresponding to the high flow threshold (higher than).}

\item{crit}{A string value corresponding to the desired performance criterion for the evaluation.}

\item{sf}{A numeric vector for specifying scaling factors if needed for correlation, variability and bias, respectively. Default values are 1-1-1.}

\item{rel_heights}{A numeric vector for specifying the relative height of each panel.}
}
\value{
A plot. Best export dimensions for default rel_heights are width = 8, height = 8.
}
\description{
Generate plot to evaluate a model by hydrological year on different subregimes
}
\examples{
data(KarstMod_dataset)
score_seq(KarstMod_dataset$date, KarstMod_dataset$QobsS + runif(365, -2, +10), KarstMod_dataset$QobsS)
}
