% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_year.R
\name{score_year}
\alias{score_year}
\title{Generate plot to evaluate a model by hydrological year on different subregimes}
\usage{
score_year(
  date,
  sim,
  obs,
  month_hyear = 9,
  lf_quantile = 0.1,
  hf_quantile = 0.9,
  crit = c("NSE", "KGE", "rpearson", "rspearman", "KGE_m", "KGE_m2", "KGENP", "LME",
    "LCE"),
  sf = c(1, 1, 1)
)
}
\arguments{
\item{date}{A date vector.}

\item{sim}{A numeric vector of simulated time series.}

\item{obs}{A numeric vector of observed time series.}

\item{month_hyear}{A numeric value corresponding to the first month of the hydrological year.}

\item{lf_quantile}{A numeric value corresponding to the low flow threshold (lower than).}

\item{hf_quantile}{A numeric value corresponding to the high flow threshold (higher than).}

\item{crit}{A string value corresponding to the desired performance criterion for the evaluation.}

\item{sf}{A numeric vector for specifying scaling factors if needed for correlation, variability and bias, respectively. Default values are 1-1-1.}
}
\value{
A plot
}
\description{
Generate plot to evaluate a model by hydrological year on different subregimes
}
\examples{
data(KarstMod_dataset)
score_year(KarstMod_dataset$date, KarstMod_dataset$QobsS + runif(365, -2, +10), KarstMod_dataset$QobsS)
}
