% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_piper_data.R
\name{transform_piper_data}
\alias{transform_piper_data}
\title{Transform analyses to coordinates for Piper diagram. This function already integrates the conversion to meq/L and to percent.}
\usage{
transform_piper_data(
  data,
  ions = c("Ca", "Mg", "Na", "K", "Cl", "SO4", "HCO3", "CO3"),
  cations = c("Ca", "Mg", "Na", "K"),
  anions = c("Cl", "SO4", "HCO3", "CO3"),
  name = NULL
)
}
\arguments{
\item{data}{A dataframe with ions as columns.}

\item{ions}{A string vector of ions to convert.}

\item{cations}{A string vector of cations.}

\item{anions}{A string vector of anions.}
}
\value{
A dataframe with three columns, including "x" and "y" which correspond to the x and y coordinates of the blank piper ggplot..
}
\description{
Transform analyses to coordinates for Piper diagram. This function already integrates the conversion to meq/L and to percent.
}
\examples{
data(pc_data)

transform_piper_data(pc_data,
                     ions = c("Ca", "Mg", "Na", "K", "SO4", "Cl", "HCO3"),
                     anions = c("Cl", "SO4", "HCO3"))
}
