% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{PET}
\alias{PET}
\title{Calculation of daily Potential Evapotranspiration (PET) with different methods.}
\usage{
PET(
  date,
  t,
  method = c("oudin2005", "hargreaves_samani1985", "turc1961", "haude1954",
    "tornthwaite_pereira2004"),
  tmax = NULL,
  tmin = NULL,
  rh = NULL,
  latitude = NULL,
  krs = 0.17,
  Rs = NULL,
  k1 = 100,
  k2 = 5,
  f_haude = c(0.26, 0.26, 0.33, 0.39, 0.39, 0.37, 0.35, 0.33, 0.31, 0.26, 0.26, 0.26),
  k = 0.69
)
}
\arguments{
\item{date}{A date vector.}

\item{t}{A numeric vector of mean daily temperature.}

\item{method}{A string value. Either "oudin2005", "hargreaves1985" or "turc1961".}

\item{tmax}{A numeric vector of maximum daily temperature.}

\item{tmin}{A numeric vector of minimum daily temperature.}

\item{rh}{A numeric vector of mean daily relative humidity.}

\item{latitude}{A numeric value. Latitude (in degrees) of the meteorological station. This is for estimating Ra.}

\item{krs}{A numeric value for estimating Rs. Radiation adjustment coefficient (generally between 0.16 and 0.19).}

\item{Rs}{A numeric vector of mean daily solar radiation.}

\item{k1}{A numeric value. Used for Oudin's method.}

\item{k2}{A numeric value. Used for Oudin's method.}

\item{f_haude}{A numeric vector of one factor for each month (12 in total). Used for Haude's method. Detailed in german \url{https://www.climate-service-center.de/products_and_publications/publications/detail/063341/index.php.de}.}

\item{k}{A numeric value corresponding to the calibration coefficient of the effective temperature. Used for Tornthwaite-Pereira's method (they recommended 0.69).}
}
\value{
A numeric vector of mean daily potentation evapotranspiration.
}
\description{
This function is a wrapper for several methods of estimation of daily PET. The arguments can be specific to one method and are often set to a default value found in the literature.
}
\details{
\itemize{
\item \insertCite{oudinWhichPotentialEvapotranspiration2005;textual}{hydros1mple}. Parameters: \emph{t}, \emph{latitude}, \emph{k1}, \emph{k2}.
\item \insertCite{hargreavesReferenceCropEvapotranspiration1985;textual}{hydros1mple}. Parameters: \emph{t}, \emph{tmin}, \emph{tmax}, \emph{latitude}.
\item \insertCite{turcEvaluationBesoinsEau1961;textual}{hydros1mple}.  Parameters: \emph{t}, \emph{tmin}, \emph{tmax}, \emph{latitude}, \emph{rh}, \emph{krs}.
\item \insertCite{thornthwaiteApproachRationalClassification1948;textual}{hydros1mple} adapted by \insertCite{pereiraAdaptationThornthwaiteScheme2004;textual}{hydros1mple}.  Parameters: \emph{t}, \emph{tmin}, \emph{tmax}, \emph{k}, \emph{latitude}.
\item \insertCite{haudeZurPraktischenBestimmung1954;textual}{hydros1mple}.  Parameters: \emph{t}, \emph{rh}, \emph{f_haude}.
}

\insertCite{xiangSimilarityDifferencePotential2020;textual}{hydros1mple} proposed a review of PET methods with a lot of details and references (which can help to understand the equation).
}
\examples{
data(KarstMod_dataset)
PET(date = KarstMod_dataset$date, t = KarstMod_dataset$T, latitude = 48)
}
\references{
{
\insertAllCited{}
}
}
